Refactor context and callables
